Enhances base URL determination for deployments
Refactors base URL determination logic in the deployment workflow.

Prioritizes the `BASE_URL` secret if defined. If not set, it checks if the `SITE_URL` is a custom domain. If so, the base URL is set to "/", otherwise it defaults to "/repo-name/".

This provides greater flexibility for deployments to custom domains and standard GitHub Pages.